+++ Assets/lilToon/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Mask by vertex id
 - Dither
+- Alpha replacement for main color 2nd/3rd
 
 ### Fixed
 - Fixed lilToonMulti convert